fix: Use root side-to-move for UCI score conversion
The -40 ELO regression was caused by using the CURRENT board side-to-move
during search instead of the ROOT side-to-move for UCI conversion.

The bug:
- During negamax search, board.sideToMove() alternates between WHITE/BLACK
- We were passing this changing value to the UCI output functions
- This caused scores to be randomly converted or not based on ply depth
- OpenBench was getting inconsistent score perspectives

The fix:
- Store rootSideToMove at the start of search (both search functions)
- Use info.rootSideToMove consistently for all UCI output
- Re-enabled the UCI White's perspective conversion

This ensures UCI output is always from White's perspective regardless
of where we are in the search tree, which is the UCI standard.

Expected: Full recovery of the -40 ELO regression

bench 19191913